tests/wpt/tests/intersection-observer/unclipped-root.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
pre, #log {
  position: absolute;
  top: 0;
  left: 200px;
}
#root {
  overflow: visible;
  height: 200px;
  width: 160px;
  border: 7px solid black;
}
#target {
  width: 100px;
  height: 100px;
  background-color: green;
}
</style>

<div id="root">
  <div id="target" style="transform: translateY(300px)"></div>
</div>

<script>
var entries = [];
var target;

runTestCycle(function() {
  target = document.getElementById("target");
  assert_true(!!target, "target exists");
  var root = document.getElementById("root");
  assert_true(!!root, "root exists");
  var observer = new IntersectionObserver(function(changes) {
    entries = entries.concat(changes)
  }, {root: root});
  observer.observe(target);
  entries = entries.concat(observer.takeRecords());
  assert_equals(entries.length, 0, "No initial notifications.");
  runTestCycle(step0, "First rAF.");
}, "Test that border bounding box is used to calculate intersection with a non-scrolling root.");

function step0() {
  target.style.transform = "translateY(195px)";
  runTestCycle(step1, "target.style.transform = 'translateY(195px)'");
  checkLastEntry(entries, 0, [15, 115, 315, 415, 0, 0, 0, 0, 8, 182, 8, 222, false]);
}

function step1() {
  target.style.transform = "";
  checkLastEntry(entries, 1, [15, 115, 210, 310, 15, 115, 210, 222, 8, 182, 8, 222, true]);
}
</script>
